requirements.txt
Flask
python-dotenv
speechrecognition
pydub
gTTS
groq
pyaudio